+++ prh-app-server/src/main/java/org/onap/dcaegen2/services/prh/tasks/KafkaConsumerTaskImpl.java
@@ -34,7 +34,7 @@ import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
 
-import javax.annotation.PostConstruct;
+import jakarta.annotation.PostConstruct;
 
 
 /**